Factored helper methods out of problem.py
