+++ glotaran/analysis/problem.py
@@ -10,6 +10,13 @@ from typing import NamedTuple
 import numpy as np
 import xarray as xr
 
+from glotaran.analysis.util import LabelAndMatrix
+from glotaran.analysis.util import find_overlap
+from glotaran.analysis.util import find_closest_index
+from glotaran.analysis.util import combine_matrices
+from glotaran.analysis.util import reduce_matrix
+from glotaran.analysis.util import calculate_matrix
+from glotaran.analysis.util import get_min_max_from_interval
 from glotaran.analysis.nnls import residual_nnls
 from glotaran.analysis.variable_projection import residual_variable_projection
 from glotaran.model import DatasetDescriptor
@@ -51,11 +58,6 @@ class GroupedProblem(NamedTuple):
 
 UngroupedBag = Dict[str, ProblemDescriptor]
 GroupedBag = Deque[GroupedProblem]
-
-
-class LabelAndMatrix(NamedTuple):
-    clp_label: list[str]
-    matrix: np.ndarray
 
 
 class Problem:
@@ -349,11 +351,11 @@ class Problem:
 
                 idx = {}
                 if weight.global_interval is not None:
-                    idx[self.model.global_dimension] = _get_min_max_from_interval(
+                    idx[self.model.global_dimension] = get_min_max_from_interval(
                         weight.global_interval, global_axis
                     )
                 if weight.model_interval is not None:
-                    idx[self.model.model_dimension] = _get_min_max_from_interval(
+                    idx[self.model.model_dimension] = get_min_max_from_interval(
                         weight.model_interval, model_axis
                     )
                 dataset.weight[idx] *= weight.value
@@ -437,7 +439,7 @@ class Problem:
         weight: xr.DataArray,
         has_scaling: bool,
     ):
-        i1, i2 = _find_overlap(self._full_axis, global_axis, atol=self._scheme.group_tolerance)
+        i1, i2 = find_overlap(self._full_axis, global_axis, atol=self._scheme.group_tolerance)
 
         for i, j in enumerate(i1):
             datasets[j].append(label)
@@ -534,7 +536,7 @@ class Problem:
         ) -> tuple[list[tuple[LabelAndMatrix, str]], float]:
             result = [
                 (
-                    _calculate_matrix(
+                    calculate_matrix(
                         self._model,
                         descriptors[problem.label],
                         problem.indices,
@@ -554,13 +556,13 @@ class Problem:
             index_results, index = results
             constraint_labels_and_matrices = list(
                 map(
-                    lambda result: _reduce_matrix(
+                    lambda result: reduce_matrix(
                         self._model, result[1], self.parameters, result[0], index
                     ),
                     index_results,
                 )
             )
-            clp, matrix = _combine_matrices(constraint_labels_and_matrices)
+            clp, matrix = combine_matrices(constraint_labels_and_matrices)
             return LabelAndMatrix(clp, matrix)
 
         results = list(
@@ -610,7 +612,7 @@ class Problem:
             descriptor = self._filled_dataset_descriptors[label]
 
             for i, index in enumerate(problem.global_axis):
-                result = _calculate_matrix(
+                result = calculate_matrix(
                     self._model,
                     descriptor,
                     {self._global_dimension: i},
@@ -622,7 +624,7 @@ class Problem:
 
                 self._clp_labels[label].append(result.clp_label)
                 self._matrices[label].append(result.matrix)
-                reduced_labels_and_matrix = _reduce_matrix(
+                reduced_labels_and_matrix = reduce_matrix(
                     self._model, label, self._parameters, result, index
                 )
                 self._reduced_clp_labels[label].append(reduced_labels_and_matrix.clp_label)
@@ -637,7 +639,7 @@ class Problem:
         self.calculate_index_independent_ungrouped_matrices()
         for group_label, group in self.groups.items():
             if group_label not in self._matrices:
-                reduced_labels_and_matrix = _combine_matrices(
+                reduced_labels_and_matrix = combine_matrices(
                     [
                         LabelAndMatrix(
                             self._reduced_clp_labels[label], self._reduced_matrices[label]
@@ -669,7 +671,7 @@ class Problem:
         for label, descriptor in self._filled_dataset_descriptors.items():
             model_axis = self._data[label].coords[self._model_dimension].values
             global_axis = self._data[label].coords[self._global_dimension].values
-            result = _calculate_matrix(
+            result = calculate_matrix(
                 self._model,
                 descriptor,
                 {},
@@ -681,7 +683,7 @@ class Problem:
 
             self._clp_labels[label] = result.clp_label
             self._matrices[label] = result.matrix
-            reduced_result = _reduce_matrix(self._model, label, self._parameters, result, None)
+            reduced_result = reduce_matrix(self._model, label, self._parameters, result, None)
             self._reduced_clp_labels[label] = reduced_result.clp_label
             self._reduced_matrices[label] = reduced_result.matrix
 
@@ -876,7 +878,7 @@ class Problem:
         for label, clp_labels in self.clp_labels.items():
 
             # find offset in the full axis
-            offset = _find_closest_index(
+            offset = find_closest_index(
                 self.data[label].coords[self._global_dimension][0].values, self._full_axis
             )
 
@@ -1161,110 +1163,3 @@ class Problem:
         )
 
         dataset[f"{name}_singular_values"] = (("singular_value_index"), v)
-
-
-def _find_overlap(a, b, rtol=1e-05, atol=1e-08):
-    ovr_a = []
-    ovr_b = []
-    start_b = 0
-    for i, ai in enumerate(a):
-        for j, bj in itertools.islice(enumerate(b), start_b, None):
-            if np.isclose(ai, bj, rtol=rtol, atol=atol, equal_nan=False):
-                ovr_a.append(i)
-                ovr_b.append(j)
-            elif bj > ai:  # (more than tolerance)
-                break  # all the rest will be farther away
-            else:  # bj < ai (more than tolerance)
-                start_b += 1  # ignore further tests of this item
-    return (ovr_a, ovr_b)
-
-
-def _calculate_matrix(
-    model: Model,
-    dataset_descriptor: DatasetDescriptor,
-    indices: dict[str, int],
-    axis: dict[str, np.ndarray],
-) -> LabelAndMatrix:
-    clp_labels = None
-    matrix = None
-
-    for scale, megacomplex in dataset_descriptor.iterate_megacomplexes():
-        this_clp_labels, this_matrix = megacomplex.calculate_matrix(
-            model, dataset_descriptor, indices, axis
-        )
-
-        if scale is not None:
-            this_matrix *= scale
-
-        if matrix is None:
-            clp_labels = this_clp_labels
-            matrix = this_matrix
-        else:
-            tmp_clp_labels = clp_labels + [c for c in this_clp_labels if c not in clp_labels]
-            tmp_matrix = np.zeros((matrix.shape[0], len(tmp_clp_labels)), dtype=np.float64)
-            for idx, label in enumerate(tmp_clp_labels):
-                if label in clp_labels:
-                    tmp_matrix[:, idx] += matrix[:, clp_labels.index(label)]
-                if label in this_clp_labels:
-                    tmp_matrix[:, idx] += this_matrix[:, this_clp_labels.index(label)]
-            clp_labels = tmp_clp_labels
-            matrix = tmp_matrix
-
-    return LabelAndMatrix(clp_labels, matrix)
-
-
-def _reduce_matrix(
-    model: Model,
-    label: str,
-    parameters: ParameterGroup,
-    result: LabelAndMatrix,
-    index: float,
-) -> LabelAndMatrix:
-    clp_labels = result.clp_label.copy()
-    if callable(model.has_matrix_constraints_function) and model.has_matrix_constraints_function():
-        clp_label, matrix = model.constrain_matrix_function(
-            label, parameters, clp_labels, result.matrix, index
-        )
-        return LabelAndMatrix(clp_label, matrix)
-    return LabelAndMatrix(clp_labels, result.matrix)
-
-
-def _combine_matrices(labels_and_matrices: list[LabelAndMatrix]) -> LabelAndMatrix:
-    masks = []
-    full_clp_labels = None
-    sizes = []
-    for label_and_matrix in labels_and_matrices:
-        (clp_label, matrix) = label_and_matrix
-        sizes.append(matrix.shape[0])
-        if full_clp_labels is None:
-            full_clp_labels = clp_label
-            masks.append([i for i, _ in enumerate(clp_label)])
-        else:
-            mask = []
-            for c in clp_label:
-                if c not in full_clp_labels:
-                    full_clp_labels.append(c)
-                mask.append(full_clp_labels.index(c))
-            masks.append(mask)
-    dim1 = np.sum(sizes)
-    dim2 = len(full_clp_labels)
-    full_matrix = np.zeros((dim1, dim2), dtype=np.float64)
-    start = 0
-    for i, m in enumerate(labels_and_matrices):
-        end = start + sizes[i]
-        full_matrix[start:end, masks[i]] = m[1]
-        start = end
-
-    return LabelAndMatrix(full_clp_labels, full_matrix)
-
-
-def _find_closest_index(index: float, axis: np.ndarray):
-    return np.abs(axis - index).argmin()
-
-
-def _get_min_max_from_interval(interval, axis):
-    minimum = np.abs(axis.values - interval[0]).argmin() if not np.isinf(interval[0]) else 0
-    maximum = (
-        np.abs(axis.values - interval[1]).argmin() + 1 if not np.isinf(interval[1]) else axis.size
-    )
-    return slice(minimum, maximum)
