+++ glotaran/builtin/models/kinetic_spectrum/test/test_coherent_artifact.py
@@ -2,7 +2,7 @@ import numpy as np
 import xarray as xr
 
 from glotaran.analysis.optimize import optimize
-from glotaran.analysis.problem import _calculate_matrix
+from glotaran.analysis.util import calculate_matrix
 from glotaran.analysis.simulation import simulate
 from glotaran.builtin.models.kinetic_spectrum import KineticSpectrumModel
 from glotaran.parameter import ParameterGroup
@@ -54,7 +54,7 @@ def test_coherent_artifact():
     time = np.asarray(np.arange(0, 50, 1.5))
 
     data = model.dataset["dataset1"].fill(model, parameters)
-    compartments, matrix = _calculate_matrix(model, data, {}, {"time": time})
+    compartments, matrix = calculate_matrix(model, data, {}, {"time": time})
 
     assert len(compartments) == 4
     for i in range(1, 4):
