+++ glotaran/builtin/models/kinetic_spectrum/test/test_spectral_constraints.py
@@ -3,7 +3,7 @@ import pytest
 import xarray as xr
 
 from glotaran.analysis.optimize import optimize
-from glotaran.analysis.problem import _calculate_matrix
+from glotaran.analysis.problem import calculate_matrix
 from glotaran.analysis.simulation import simulate
 from glotaran.builtin.models.kinetic_spectrum import KineticSpectrumModel
 from glotaran.builtin.models.kinetic_spectrum.spectral_constraints import (
@@ -61,7 +61,7 @@ def test_spectral_constraint():
 
     time = np.asarray(np.arange(0, 50, 1.5))
     dataset = model.dataset["dataset1"].fill(model, wanted_parameters)
-    compartments, matrix = _calculate_matrix(model, dataset, {}, {"time": time})
+    compartments, matrix = calculate_matrix(model, dataset, {}, {"time": time})
 
     assert len(compartments) == 2
     assert matrix.shape == (time.size, 2)
